chore: 이슈 정보 생성 코드 분리
