--- app/v2/middleware/is_content_created.py
@@ -1,1 +1,0 @@
-
